Convert CLI sessions help tests as unit tests
